refactor(export): remove Export tab, move exports to header for both modes
TopBar now handles Ethernet and FC exports directly — CSV, PDF, and Print
buttons in the header work for both modes. The dedicated Export tab is
removed as it duplicated functionality already present in the header.
