+++ src/components/TopBar.tsx
@@ -12,9 +12,12 @@ import {
   TooltipTrigger,
 } from '@/components/ui/tooltip'
 import { useResultStore } from '@/store/resultStore'
+import { useFCResultStore } from '@/store/fcResultStore'
 import { downloadBomCsv } from '@/features/export/exportCsv'
+import { downloadFCBomCsv } from '@/features/export/exportFCCsv'
 import { generatePdfBlob } from '@/features/export/exportPdf'
-import { getLastTopologyPng } from '@/features/topology'
+import { generateFCPdfBlob } from '@/features/export/exportFCPdf'
+import { getLastTopologyPng, getLastFCTopologyPng } from '@/features/topology'
 
 interface TopBarProps {
   mode: 'ethernet' | 'fc'
@@ -24,28 +27,51 @@ interface TopBarProps {
 export function TopBar({ mode, onModeChange }: TopBarProps) {
   const { t } = useTranslation()
   const bom = useResultStore(useShallow((s) => s.bom))
+  const fcBom = useFCResultStore(useShallow((s) => s.bom))
+  const activeBom = mode === 'fc' ? fcBom : bom
   const [pdfGenerating, setPdfGenerating] = useState(false)
 
   const handleCsv = () => {
-    if (bom) downloadBomCsv(bom)
+    if (mode === 'fc' && fcBom) downloadFCBomCsv(fcBom)
+    else if (bom) downloadBomCsv(bom)
   }
 
   const handlePdf = async () => {
-    if (!bom) return
-    setPdfGenerating(true)
-    try {
-      const png = getLastTopologyPng()
-      const blob = await generatePdfBlob(bom, png ?? undefined)
-      const url = URL.createObjectURL(blob)
-      const a = document.createElement('a')
-      a.href = url
-      a.download = 'netstack-report.pdf'
-      a.click()
-      URL.revokeObjectURL(url)
-    } catch {
-      // Silently fail — user can retry
-    } finally {
-      setPdfGenerating(false)
+    if (mode === 'fc') {
+      if (!fcBom) return
+      setPdfGenerating(true)
+      try {
+        const pngA = getLastFCTopologyPng('A') ?? undefined
+        const pngB = getLastFCTopologyPng('B') ?? undefined
+        const blob = await generateFCPdfBlob(fcBom, pngA, pngB)
+        const url = URL.createObjectURL(blob)
+        const a = document.createElement('a')
+        a.href = url
+        a.download = 'netstack-fc-report.pdf'
+        a.click()
+        URL.revokeObjectURL(url)
+      } catch {
+        // Silently fail — user can retry
+      } finally {
+        setPdfGenerating(false)
+      }
+    } else {
+      if (!bom) return
+      setPdfGenerating(true)
+      try {
+        const png = getLastTopologyPng()
+        const blob = await generatePdfBlob(bom, png ?? undefined)
+        const url = URL.createObjectURL(blob)
+        const a = document.createElement('a')
+        a.href = url
+        a.download = 'netstack-report.pdf'
+        a.click()
+        URL.revokeObjectURL(url)
+      } catch {
+        // Silently fail — user can retry
+      } finally {
+        setPdfGenerating(false)
+      }
     }
   }
 
@@ -69,7 +95,7 @@ export function TopBar({ mode, onModeChange }: TopBarProps) {
               size="icon"
               className="h-9 w-9"
               onClick={handleCsv}
-              disabled={!bom}
+              disabled={!activeBom}
             >
               <FileSpreadsheet className="h-4 w-4" />
             </Button>
@@ -84,7 +110,7 @@ export function TopBar({ mode, onModeChange }: TopBarProps) {
               size="icon"
               className="h-9 w-9"
               onClick={handlePdf}
-              disabled={!bom || pdfGenerating}
+              disabled={!activeBom || pdfGenerating}
             >
               {pdfGenerating ? (
                 <Loader2 className="h-4 w-4 animate-spin" />
