+++ src/features/export/index.ts
@@ -1,3 +1,2 @@
-export { ExportTab } from './ExportTab'
 export { downloadFCBomCsv, buildFCCsvString } from './exportFCCsv'
 export { generateFCPdfBlob } from './exportFCPdf'
